Create Encrytion.py
Write a program that uses a dictionary to assign "codes" to each letter of the alphabet.

For example:
 codes = { 'A' : '%', 'a' : '9', 'B' : '@', 'b' : '#', etc...}

Using this example, the letter A would be assigned the symbol %, the letter a would be assigned the number 9, the letter B would be assigned the symbol @, and so forth.

The program should open a specified text file, read its contents, and then use the dictionary to write an encrypted version of the file's contents to a second file. Each character in the second file should contain the code for the corresponding character in the first file.